+++ src/onegov.election_day/onegov/election_day/path.py
@@ -5,6 +5,7 @@ from onegov.ballot import Vote, VoteCollection
 from onegov.election_day import ElectionDayApp
 from onegov.election_day.models import Principal
 from onegov.election_day.collections import ArchivedResultCollection
+from onegov.election_day.collections import SubscriberCollection
 from onegov.user import Auth
 
 
@@ -26,6 +27,11 @@ def get_manage_elections(app, page=0):
 @ElectionDayApp.path(model=VoteCollection, path='/manage/votes')
 def get_manage_votes(app, page=0):
     return VoteCollection(app.session(), page=page)
+
+
+@ElectionDayApp.path(model=SubscriberCollection, path='/manage/subscribers')
+def get_manage_subscribers(app):
+    return SubscriberCollection(app.session())
 
 
 @ElectionDayApp.path(model=Election, path='/election/{id}')